move write_meta
