+++ ydb/core/persqueue/pqtablet/partition/partition_compactification.cpp
@@ -1,7 +1,7 @@
 #include "partition.h"
 #include "partition_compactification.h"
 #include <ydb/core/persqueue/pqtablet/common/logging.h>
-#include <ydb/core/persqueue/write_meta.h>
+#include <ydb/core/persqueue/public/write_meta/write_meta.h>
 #include "partition_util.h"
 
 namespace NKikimr::NPQ {
