+++ ydb/core/persqueue/public/write_meta/write_meta.h
