+++ ydb/core/persqueue/public/write_meta/ya.make
@@ -1,0 +1,14 @@
+LIBRARY()
+
+SRCS(
+    write_meta.cpp
+)
+
+PEERDIR(
+    ydb/core/persqueue/public
+    ydb/public/sdk/cpp/src/client/persqueue_public
+    ydb/public/sdk/cpp/src/client/topic
+)
+
+END()
+
