+++ ydb/services/deprecated/persqueue_v0/grpc_pq_write_actor.cpp
@@ -4,7 +4,7 @@
 #include <ydb/core/persqueue/writer/metadata_initializers.h>
 
 #include <ydb/core/persqueue/public/pq_database.h>
-#include <ydb/core/persqueue/write_meta.h>
+#include <ydb/core/persqueue/public/write_meta/write_meta.h>
 #include <ydb/library/services/services.pb.h>
 #include <ydb/public/lib/deprecated/kicli/kicli.h>
 #include <ydb/library/persqueue/topic_parser/topic_parser.h>
