+++ ydb/services/persqueue_v1/actors/partition_actor.cpp
@@ -3,7 +3,7 @@
 
 #include <limits>
 #include <ydb/core/persqueue/codecs/pqv1.h>
-#include <ydb/core/persqueue/write_meta.h>
+#include <ydb/core/persqueue/public/write_meta/write_meta.h>
 #include <ydb/core/persqueue/writer/source_id_encoding.h>
 
 #include <ydb/public/api/protos/ydb_persqueue_v1.pb.h>
